Open the door
A door can be opened only if two or more magical items are available. If there is only one magical item is available, then a jewel can be added to it to open the door. The magical items are represented as even numbers and jewels are represented as negative numbers. The program must accept three integers as the input and print YES if the door can be opened. Else the program must print NO.

Example Input/Output 1:
Input:
24 25 64

Output:
YES

Example Input/Output 2:
Input:
53 79 -2

Output:
NO